+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## [0.14.0](https://github.com/paulkre/bevy_image_export/compare/v0.13.0...v0.14.0) (2025-10-19)
 
 ### Features